refactor form-chatbot into focused modules and intercept before-submit to avoid failure message
